refactor: extrair checkGameBtn e selectNewGameBtn p/ components
