+++ src/components/GameHeader/components/CheckGameButton/style.css
@@ -1,0 +1,3 @@
+#check-btn:hover {
+    cursor: pointer;
+}
