+++ src/components/GameHeader/components/CheckGameButton/style.ts
@@ -1,0 +1,10 @@
+import { CSSProperties } from 'react';
+
+export const checkButtonStyle: CSSProperties = {
+    backgroundColor: 'white',
+    border: '1px solid rgba(110,110,110,0.5)',
+    padding: '5px',
+    textDecoration: 'none',
+    borderRadius: '5px',
+    fontSize: '16px',
+};
