+++ src/components/GameHeader/components/SelectNewGameButton/style.css
@@ -1,7 +1,3 @@
-#check-btn:hover {
-    cursor: pointer;
-}
-
 button#btn-newgame:hover {
     cursor: pointer;
 }
