+++ src/components/GameHeader/components/index.ts
@@ -1,0 +1,2 @@
+export * from './CheckGameButton';
+export * from './SelectNewGameButton';
